requirements.txt
torch>=2.0.0
diffusers == 0.31.0
transformers>=4.35.0
accelerate>=0.27.0
safetensors>=0.4.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tqdm>=4.66.0
huggingface-hub>=0.19.0
bitsandbytes>=0.41.0
optimum-quanto
multiformats